requirements.txt
asgiref==3.5.0
autopep8==1.6.0
cffi==1.15.0
cryptography==37.0.1
Django==4.0.4
django-extensions==3.1.5
django-multiselectfield==0.1.12
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
mysqlclient==2.1.0
Pillow==9.1.0
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
Werkzeug==2.1.2
whitenoise==6.0.0